requirements.txt
numpy==1.17.1
pytorch-nlp==0.4.1
PyYAML==5.1.2
sacremoses==0.0.33
scikit-learn==0.21.3
sklearn==0.0
torch==1.1.0
torchtext==0.4.0
tqdm==4.35.0
